# Replace boxing Stacks in BamlReader, decrease allocs, improve performance