## Description

Replaces boxing `Stack` with `Stack<bool>`, increasing overall performance and decreasing allocations/code-size.
Also replaces non-generic `Stack` with `Stack<BamlNodeInfo>` and makes it `readonly`.

For ilustration, here are some benchmarks:

### 1 boxing stack vs. generic

| Method       | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| BoxingStack  |  67.83 ns |   1.107 ns |    0.981 ns | 0.0238 |         869 B |         400 B |
| GenericStack |  22.23 ns |   0.213 ns |    0.189 ns | 0.0038 |         210 B |          64 B |

<details><summary>Benchmark code</summary>

```c#
[Benchmark]
public void BoxingStack()
{
    Stack stack = new Stack();
    stack.Push(true);

    int i = 0;
    while (i < 10)
    {
        if ((bool)stack.Peek())
        {
            stack.Pop();
            stack.Push(true);
        }
        i++;
    }


}
[Benchmark]
public void GenericStack()
{
    Stack<bool> stack = new Stack<bool>();
    stack.Push(true);

    int i = 0;
    while (i < 10)
    {
        if (stack.Peek())
        {
            stack.Pop();
            stack.Push(true);
        }
        i++;
    }

```

</details>

### 3 boxing stacks vs. generic

| Method       | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|------------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| BoxingStack  | 207.90 ns |   3.092 ns |    2.582 ns | 0.0715 |       1,798 B |        1200 B |
| GenericStack |  59.11 ns |   0.923 ns |    0.863 ns | 0.0114 |         441 B |         192 B |

<details><summary>Benchmark code</summary>

```c#
[Benchmark]
public void BoxingStack()
{
    Stack stack = new Stack();
    stack.Push(true);
    Stack stack2 = new Stack();
    stack2.Push(true);
    Stack stack3 = new Stack();
    stack3.Push(true);

    int i = 0;
    while (i < 10)
    {
        if ((bool)stack.Peek())
        {
            stack.Pop();
            stack.Push(true);
        }
        if ((bool)stack2.Peek())
        {
            stack2.Pop();
            stack2.Push(true);
        }
        if ((bool)stack3.Peek())
        {
            stack3.Pop();
            stack3.Push(true);
        }
        i++;
    }


}
[Benchmark]
public void GenericStack()
{
    Stack<bool> stack = new Stack<bool>();
    stack.Push(true);
    Stack<bool> stack2 = new Stack<bool>();
    stack2.Push(true);
    Stack<bool> stack3 = new Stack<bool>();
    stack3.Push(true);

    int i = 0;
    while (i < 10)
    {
        if (stack.Peek())
        {
            stack.Pop();
            stack.Push(true);
        }
        if (stack2.Peek())
        {
            stack2.Pop();
            stack2.Push(true);
        }
        if (stack3.Peek())
        {
            stack3.Pop();
            stack3.Push(true);
        }
        i++;
    }

```

</details>

## Customer Impact

Increased performance, decreased allocations.

## Regression

No.

## Testing

Local build, app run.

## Risk

Low.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9397)